+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/RepositoryFacade.kt
@@ -9,6 +9,7 @@ import io.getstream.chat.android.livedata.extensions.lastMessage
 import io.getstream.chat.android.livedata.extensions.users
 import io.getstream.chat.android.livedata.model.ChannelConfig
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepository
+import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionRepository
 import io.getstream.chat.android.livedata.request.AnyChannelPaginationRequest
 import io.getstream.chat.android.livedata.request.isRequestingMoreThanLastMessage
 import kotlinx.coroutines.CoroutineScope
