+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/builder/RepositoryFactory.kt
@@ -8,8 +8,6 @@ import io.getstream.chat.android.livedata.repository.ChannelRepository
 import io.getstream.chat.android.livedata.repository.ChannelRepositoryImpl
 import io.getstream.chat.android.livedata.repository.QueryChannelsRepository
 import io.getstream.chat.android.livedata.repository.QueryChannelsRepositoryImpl
-import io.getstream.chat.android.livedata.repository.ReactionRepository
-import io.getstream.chat.android.livedata.repository.ReactionRepositoryImpl
 import io.getstream.chat.android.livedata.repository.SyncStateRepository
 import io.getstream.chat.android.livedata.repository.SyncStateRepositoryImpl
 import io.getstream.chat.android.livedata.repository.UserRepository
@@ -17,6 +15,8 @@ import io.getstream.chat.android.livedata.repository.UserRepositoryImpl
 import io.getstream.chat.android.livedata.repository.database.ChatDatabase
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepository
 import io.getstream.chat.android.livedata.repository.domain.message.MessageRepositoryImpl
+import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionRepository
+import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionRepositoryImpl
 
 internal class RepositoryFactory(
     private val database: ChatDatabase,
