+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/livedata/repository/database/ChatDatabase.kt
@@ -9,7 +9,6 @@ import androidx.sqlite.db.SupportSQLiteDatabase
 import io.getstream.chat.android.livedata.dao.ChannelConfigDao
 import io.getstream.chat.android.livedata.dao.ChannelDao
 import io.getstream.chat.android.livedata.dao.QueryChannelsDao
-import io.getstream.chat.android.livedata.dao.ReactionDao
 import io.getstream.chat.android.livedata.dao.SyncStateDao
 import io.getstream.chat.android.livedata.dao.UserDao
 import io.getstream.chat.android.livedata.entity.AttachmentEntity
@@ -18,7 +17,6 @@ import io.getstream.chat.android.livedata.entity.ChannelEntity
 import io.getstream.chat.android.livedata.entity.ChannelSortInnerEntity
 import io.getstream.chat.android.livedata.entity.CommandInnerEntity
 import io.getstream.chat.android.livedata.entity.QueryChannelsEntity
-import io.getstream.chat.android.livedata.entity.ReactionEntity
 import io.getstream.chat.android.livedata.entity.SyncStateEntity
 import io.getstream.chat.android.livedata.entity.UserEntity
 import io.getstream.chat.android.livedata.repository.database.converter.DateConverter
@@ -30,6 +28,8 @@ import io.getstream.chat.android.livedata.repository.database.converter.SetConve
 import io.getstream.chat.android.livedata.repository.database.converter.SyncStatusConverter
 import io.getstream.chat.android.livedata.repository.domain.message.MessageDao
 import io.getstream.chat.android.livedata.repository.domain.message.MessageInnerEntity
+import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionDao
+import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionEntity
 
 @Database(
     entities = [
