+++ stream-chat-android-offline/src/test/java/io/getstream/chat/android/livedata/Mother.kt
@@ -15,10 +15,10 @@ import io.getstream.chat.android.client.models.Reaction
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.client.utils.SyncStatus
 import io.getstream.chat.android.livedata.entity.AttachmentEntity
-import io.getstream.chat.android.livedata.entity.ReactionEntity
 import io.getstream.chat.android.livedata.entity.UserEntity
 import io.getstream.chat.android.livedata.repository.domain.message.MessageEntity
 import io.getstream.chat.android.livedata.repository.domain.message.MessageInnerEntity
+import io.getstream.chat.android.livedata.repository.domain.reaction.ReactionEntity
 import io.getstream.chat.android.test.positiveRandomInt
 import io.getstream.chat.android.test.randomBoolean
 import io.getstream.chat.android.test.randomCID
